docs : Common에 들어갈 object를 util package로 옮김
